+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Allow services to specify services / tasks that should always be deployed "after" them, but should not constitute a direct dependency.
+
 ### Changed
 - `daemon restart` will start the daemon if it is not running